alert when delete row

+++ client/src/components/Admin/AdminCategory/CategoryTable.tsx
@@ -1,6 +1,8 @@
 import React, { useEffect, useState } from "react";
 import { AdminTable } from "../Content/AdminTable";
 import { AdminButton } from "../Content/AdminButton";
+import Swal from 'sweetalert2';
+
 
 export const CategoryTable = () => {
   const columns = [
@@ -37,22 +39,46 @@ export const CategoryTable = () => {
     });
   };
   const deleteTableRow = async (row: any) => {
-    const body = { id: row.id };
-    const res = await fetch("/api/category", {
-      method: "DELETE",
-      mode: "cors",
-      headers: {
-        "Content-Type": "application/json"
-      },
-      body: JSON.stringify(body)
-    });
-    if (res.status === 200) {
-      const { status } = await res.json();
-      console.log(status);
-    } else {
-      const { errors } = await res.json();
-      console.log(errors);
-    }
+    Swal.fire({
+      title: `<p class="text-admin-title">Are you sure?</p>`,
+      html: `<p class="text-admin-title">You gonna delete row id ${row.id}!</p>`,
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#BA3632',
+      cancelButtonColor: '#2799E9',
+      confirmButtonText: 'Yes, delete it!',
+      background: '#1E2A31',
+    }).then(async (result) => {
+      if (result.value) {
+        const body = { id: row.id };
+        const res = await fetch("/api/category", {
+          method: "DELETE",
+          mode: "cors",
+          headers: {
+            "Content-Type": "application/json"
+          },
+          body: JSON.stringify(body)
+        });
+        if (res.status === 200) {
+          const { status } = await res.json();
+          Swal.fire({
+            title: '<p class="text-admin-title">DELETED</p>',
+            html: `<p class="text-admin-title">${status}</p>`,
+            background: '#1E2A31',
+            confirmButtonColor: '#2799E9'
+          })
+          getTableData();
+        } else {
+          const { errors } = await res.json();
+          Swal.fire({
+            title: '<p class="text-admin-title">ERROR</p>',
+            html: `<p class="text-admin-title">${errors[0].message}</p>`,
+            background: '#1E2A31',
+            confirmButtonColor: '#2799E9'
+          })
+        }
+      }
+    })
   }
   useEffect(() => {
     getTableData();
